Updated typeshed stubs to the latest version.

+++ packages/pyright-internal/typeshed-fallback/stubs/Pillow/PIL/ImageFont.pyi
@@ -44,7 +44,7 @@ class FreeTypeFont:
         direction: Literal["ltr", "rtl", "ttb"] | None = None,
         features: Incomplete | None = None,
         language: str | None = None,
-    ) -> int: ...
+    ) -> float: ...
     def getbbox(
         self,
         text: str | bytes,
